Add files via upload
Credit score cards uses personal information and data submitted by credit card applicants as well as historical data and try to decide whether to issue a credit card to the applicant. Using the 1st dataset which is a filled in form and the bank’s data, we are trying to predict the probability of future defaults and credit card borrowings.